refactor: comprehensive backend/frontend restructuring with new feature modules
Backend:
- Refactor auth module: organize AuthController and AuthService into dedicated directories (controller/service)
- Add new DashboardController with service layer
- Enhance security: improve JwtAuthFilter and JwtService implementations
- Extend bid system: update BidService, repository, DTOs with additional fields
- Expand contract management: add Milestone domain, MilestoneStatus enum, and related services
- Enhance project module: extend ProjectService, DTOs (filter, summary responses), and repository queries
- Add payment system foundation: new StripeConfig, payment DTOs, and database migrations
- Implement notification system: new notification module with database schema
- Add user profile service: ProfileService for user profile management
- Email verification: add email verification migration and requests (ForgotPasswordRequest, ResetPasswordRequest)
- Database: add 7 new migrations (V8-V14) for payment, notifications, profile pictures, email verification
- Update dependencies: modify pom.xml for payment/notification support

Frontend:
- Refactor auth module: modernize components (LoginForm, RegisterForm, SignupFlow), migrate auth-store location
- Restructure component hierarchy: migrate from components/layout to internal modules
- Remove legacy utilities: delete axios setup, mock data, old auth error handlers
- Build dashboard module: new ClientDashboard and FreelancerDashboard pages with layout
- Implement bid system UI: new bid module with forms, cards, and bid management pages
- Expand contract UI: new contract detail pages, milestone components, confirmation dialogs
- Extend project pages: client project creation/editing, bid panels, improved project listing
- Add payment module: payment form components and payment page
- Implement notification UI: notification module with service integration
- Build profile management: profile pages and edit components for both roles
- Add settings page: user settings dashboard
- Create milestone UI: milestone cards and management components
- Implement review system: review cards and review forms
- Migrate navigation: update Sidebar, Topbar, and create MobileSidebar in modules/navigation
- Landing page overhaul: update all hero, features, testimonials, and role selection components
- Update routing: add forgot-password, reset-password, verify-email flows
- Enhance styling: update globals.css, components.json, postcss config
- Update TypeScript: modify tsconfig for new module structure
- Lock file: update pnpm-lock.yaml for dependency changes

This commit consolidates major refactoring efforts establishing cleaner architecture,
improved separation of concerns, and foundation for payment/notification features.

+++ backend/src/main/java/com/freelancehub/freelancehub/bid/domain/Bid.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @SuppressWarnings("FieldMayBeFinal")
@@ -37,7 +38,7 @@ public class Bid {
     private Project project;
 
     @Column(nullable = false)
-    private double proposedPrice;
+    private BigDecimal proposedPrice;
 
     @Column(nullable = false, columnDefinition = "TEXT")
     private String coverLetter;
