+++ backend/src/main/java/com/freelancehub/freelancehub/bid/dto/CreateBidRequest.java
@@ -2,10 +2,12 @@ package com.freelancehub.freelancehub.bid.dto;
 
 import jakarta.validation.constraints.*;
 
+import java.math.BigDecimal;
+
 public record CreateBidRequest(
 
         @Positive(message = "Proposed price must be positive")
-        double proposedPrice,
+        BigDecimal proposedPrice,
 
         @NotBlank(message = "Cover letter is required")
         @Size(min = 50, message = "Cover letter must be at least 50 characters")
