+++ backend/src/main/java/com/freelancehub/freelancehub/project/domain/Project.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -31,10 +32,10 @@ public class Project {
     private String description;
 
     @Column(nullable = false)
-    private double budgetMin;
+    private BigDecimal budgetMin;
 
     @Column(nullable = false)
-    private double budgetMax;
+    private BigDecimal budgetMax;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
