+++ backend/src/main/java/com/freelancehub/freelancehub/project/dto/ProjectFilter.java
@@ -3,11 +3,13 @@ package com.freelancehub.freelancehub.project.dto;
 import com.freelancehub.freelancehub.project.domain.ProjectCategory;
 import com.freelancehub.freelancehub.project.domain.ProjectStatus;
 
+import java.math.BigDecimal;
+
 // Query params for filtering — all optional
 public record ProjectFilter(
         ProjectCategory category,
         ProjectStatus status,
-        Double budgetMin,
-        Double budgetMax,
+        BigDecimal budgetMin,
+        BigDecimal budgetMax,
         String skill     // filters projects that require this skill
 ) {}
