+++ backend/src/main/java/com/freelancehub/freelancehub/project/dto/ProjectResponse.java
@@ -3,6 +3,7 @@ package com.freelancehub.freelancehub.project.dto;
 import com.freelancehub.freelancehub.project.domain.ProjectCategory;
 import com.freelancehub.freelancehub.project.domain.ProjectStatus;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -12,8 +13,8 @@ public record ProjectResponse(
         String id,
         String title,
         String description,
-        double budgetMin,
-        double budgetMax,
+        BigDecimal budgetMin,
+        BigDecimal budgetMax,
         ProjectStatus status,
         ProjectCategory category,
         List<String> requiredSkills,
